+++ src/styles/globals.css
@@ -4,7 +4,7 @@
 
 body {
   /*테스트용 배경이에요*/
-  background-color: #fff;
+  background-color: "#0F172A";
 }
 
 button,
@@ -27,6 +27,6 @@ span::-webkit-scrollbar {
   }
 
   .modal-container {
-    @apply fixed left-1/2 top-1/2 z-30 -translate-x-1/2 -translate-y-1/2 bg-[#1E293B] rounded-12;
+    @apply fixed left-1/2 top-1/2 z-30 -translate-x-1/2 -translate-y-1/2 rounded-12 bg-[#1E293B];
   }
 }
